Aplicando Composição no método constructor para evitar repetição de código

+++ src/Autenticavel.java
@@ -1,5 +1,5 @@
 
 public abstract interface Autenticavel {
 	public  abstract void setSenha(int senha);
-	public abstract boolean autentica(double senha);
+	public abstract boolean autentica(int senha);
 }
